Do not redefine stack size for threads
It creates unnecessary complexity with
stack usage with minimal memory usage winning

+++ sources/include/machinarium/thread.h
@@ -18,7 +18,7 @@ struct mm_thread {
 	void *arg;
 };
 
-int mm_thread_create(mm_thread_t *, int, mm_thread_function_t, void *);
+int mm_thread_create(mm_thread_t *, size_t, mm_thread_function_t, void *);
 int mm_thread_join(mm_thread_t *);
 int mm_thread_set_name(char *);
 int mm_thread_disable_cancel(void);
